app/requirements.txt
awsebcli==3.18.1
boto==2.49.0
boto3==1.13.6
botocore==1.15.49
cachetools==4.1.0
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi
coreapi-cli
cryptography==2.9.2
defusedxml==0.7.0rc1
Django
django-autofixture
django-cors-headers
django-storages==1.9.1
djangorestframework
docutils==0.15.2
future==0.16.0
httplib2==0.18.1
idna==2.7
jmespath==0.9.5
oauthlib==3.1.0
pathspec==0.5.9
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.6
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.24.3
viberbot==1.0.11
wcwidth==0.1.9